Merge changes from previous repo. (See Detailed Comment).
Oof. Okay, so, initially, I created this theme as a modified clone of the quickstart theme instead of forking Persona. This was the best way I could find to merge my changes to the Persona theme with the existing repo history, since I don't really want authored commits containing work I didn't actually do. So as a result we have this monstrous single commit containing all the changes I've made so far.